Workaround for bug in 'post'-script with localized installer messages.
The way to detect whether a package install was rejected because there is a newer version installed was severely broken on localized (non-english) operating systems and resulted in an install error.
This is because it was based on grepping for an english language error message from '/usr/sbin/installer'.

As I have not found any way to get installer to display error messages in a requested language/locale (it always seems to go by the AppleLanguages setting in /Library/Preferences/.GlobalPreferences.plist), I was instead forced to gather all the possible error messages for this particular key in different versions of Mac OS X and hardcode it into the 'post' script.

The error messages were all taken from the key TARGET_VOLUME_VERS_CONFLICT in Distribution.strings present in all localizations of /System/Library/PrivateFrameworks/Install.framework.